pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evoskill"
version = "0.1.0"
description = "会造工具的 AI 对话系统 —— 越用越聪明，越用越强大"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "EvoSkill Team"},
]
keywords = ["ai", "agent", "skills", "llm", "autonomous"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "structlog>=24.0.0",
    "aiofiles>=23.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "python-frontmatter>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
vectorstore = ["chromadb>=0.5.0"]

# uv 依赖组配置
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
]
lint = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]
all = [
    {include-group = "dev"},
    "chromadb>=0.5.0",
]

[project.scripts]
evoskill = "evoskill.cli.main:app"
esk = "evoskill.cli.main:app"

[project.urls]
Homepage = "https://github.com/evoskill/evoskill"
Documentation = "https://docs.evoskill.ai"
Repository = "https://github.com/evoskill/evoskill"
Issues = "https://github.com/evoskill/evoskill/issues"

[tool.hatch.build.targets.wheel]
packages = ["evoskill"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=evoskill --cov-report=term-missing"
